requirements.txt
python-telegram-bot
pytesseract
Pillow
python-dotenv